task 7_3

+++ 7_lesson/Homework/7_3/Program.cs
@@ -31,7 +31,7 @@ void Print(int[,] arr)
     Console.WriteLine();
 }
 
-int Average(int[,] mas)
+double Average(double[,] mas)
 {
     
 }
